+++ R/helpers_pairwise_comparison.R
@@ -313,15 +313,17 @@ pairwise_p <- function(data,
 
       # safeguarding against edge cases
       aovmodel$model %<>%
-        dplyr::mutate(.data = .,
-                      x = forcats::fct_relabel(
-                        .f = x,
-                        .fun =  ~ stringr::str_replace(
-                          string = .x,
-                          pattern = "-",
-                          replacement = "_"
-                        )
-                      ))
+        dplyr::mutate(
+          .data = .,
+          x = forcats::fct_relabel(
+            .f = x,
+            .fun = ~ stringr::str_replace(
+              string = .x,
+              pattern = "-",
+              replacement = "_"
+            )
+          )
+        )
 
       # extracting and cleaning up Tukey's HSD output
       df_tukey <- stats::TukeyHSD(x = aovmodel, conf.level = 0.95) %>%